Fix error in browser console on every tab close
Whenever the currently focused tab is closed, both the onRemoved and
onActivated events fire at the same time, because a new tab becomes
active.

Our handler for onActivated loops through an array right as onRemoved
removes an element from that array. This leads to a javascript error
which gets printed in the browser console.

I've added a simple check to make sure the array entry exists before
using it which prevents the error, but I feel like a larger rework
should happen to avoid the race entirely.